S7.8.5_A3.1_T7 through S7.8.5_A3.1_T9 were invalid as written.  The only way to specify unicode flags to literal regexp patterns is through eval and similar mechanisms (e.g., Function constructor).  Fixed.
All but two of the remaining disabled tests in excludelist.xml were disabled because the tests polluted
the global JS environment, not because the tests themselves were invalid.  This is no longer an issue as
of the latest test262 harness (actually it wasn't an issue about a year ago either).  Re-enabled.