Bug 1128: use Date.prototype.getFullYear() in S15.2.2.1_A2_T5 Bug 1131: use String.prototype.charAt() in 15.12.3-11-26
Both of these changes were accidentaly dropped when I regenerated patches for these bugs. I have grepped the testsuite to ensure that the annexB functions only appear in the annexB section of the testsuite.

NOTE: Also regenerated the website based on this change. (Still waiting for permission to actually update the live website)