Console Runner: Fix Race Condition
When spawning threads from within a `for` loop, the original
parallelization implementation relied on a function closure to access
each `TestCase` instance.  Due to the concurrent and non-deterministic
nature of thread processing, the value of this reference at execution
time was unstable.

Explicitly specify the `TestCase` instance to the `Thread` constructor,
ensuring that each thread is run with the intended value, regardless of
execution order.