Install linter dependencies in CI
Ensure that the linter's Python dependencies are satisfied in the
continuous integration environment.

Note that this is not currently necessary but only because the linter's
dependencies happen to align with those of the test generation tool. The
dependencies of the two tool should be satisfied explicitly in order to
promote isolation.