AsyncFunction: Add tests ensuring the non-native promises behaviour
This commit adds 1 more tests ensuring the optimized behaviour of await
(see tc39/ecma262#1250) in the following cases:
- awaiting on a non-promise, non-thenable object

It also renames the previous test for non-promise (a "thenable" object)
to distinguish from the new case.

The commit adds checks for proper await/promises interleaving in the
aforementioned cases and includes a small code clean-up.