CI: Don't mark CI as broken if a new test fails.
In general, the CI job should only be marked as broken (with a red cross) if
the result should block the PR from landing (or if the result should at least
be carefully reviewed before landing).

Having failing tests in some implementations, on the other hand, does not block
the PR. On the contrary, it means the test is valuable, because it proves the
presence of a bug in the implementation.

The current setup is unnecessarily confusing, particularly for newcomers to the
project.